Move pkg event -> api/storageapi/eventsender
